refactor: clean up the search widget sources
